Organizing the FullDaysCalculator for Nebraska
