+++ spec/models/child_spec.rb
@@ -153,13 +153,13 @@ RSpec.describe Child, type: :model do
     end
 
     describe '#nebraska_full_days' do
-      let(:calculator_class) { class_double(NebraskaFullDaysCalculator, new: nil) }
-      let(:calculator_instance) { instance_double(NebraskaFullDaysCalculator, call: nil) }
+      let(:calculator_class) { class_double(Nebraska::FullDaysCalculator, new: nil) }
+      let(:calculator_instance) { instance_double(Nebraska::FullDaysCalculator, call: nil) }
 
       context 'when using live algorithms' do
-        it 'calls the NebraskaFullDaysCalculator service' do
+        it 'calls the Nebraska::FullDaysCalculator service' do
           allow(Rails.application.config).to receive(:ff_ne_live_algorithms).and_return(true)
-          allow(NebraskaFullDaysCalculator)
+          allow(Nebraska::FullDaysCalculator)
             .to receive(:new).with(child: child,
                                    date: Time.current.to_date,
                                    scope: :for_month).and_return(calculator_instance)
@@ -169,7 +169,7 @@ RSpec.describe Child, type: :model do
       end
 
       context 'when using temporary dashboard values' do
-        it 'does not call the NebraskaFullDaysCalculator service' do
+        it 'does not call the Nebraska::FullDaysCalculator service' do
           allow(Rails.application.config).to receive(:ff_ne_live_algorithms).and_return(false)
           child.nebraska_full_days(Time.current.to_date)
           expect(calculator_class).not_to have_received(:new)
